Add failing tests for previousAttributes and…
…rollbackAttributes

Current behavior is to stash an attribute’s value on EVERY call to .set(). The intent seems to be to rollback to the value fetched from the persistence layer.

As illustrated by the tests:

```
post.get(‘excerpt’); // ‘Was a gambler.’

post.set('excerpt', 'Became the sheriff.');
post.set('excerpt', 'Became the chief.');
post.set('excerpt', 'Became the mayor.');

post.get('previousAttributes.excerpt’); // 'Became the chief.'
```

The previous value is simply the previous change ('Became the chief’) instead of the “original” value (“Was a gambler”).

The new tests FAIL to show that “previousAttributes” and “rollbackAttributes” to not behave as expected.